// src/lib/spacetime-db/resources/resource-tags.ts
/**
 * Resource tag definitions based on resource_desc.json data
 */

export enum ResourceTag {
  Tree = 'Tree',
  WoodLogs = 'Wood Logs',
  Sapling = 'Sapling',
  Stick = 'Stick',
  Stump = 'Stump',
  Berry = 'Berry',
  Fruit = 'Fruit',
  Flower = 'Flower',
  Mushroom = 'Mushroom',
  FiberPlant = 'Fiber Plant',
  WildGrain = 'Wild Grain',
  WildVegetable = 'Wild Vegetable',
  Insects = 'Insects',
  Rock = 'Rock',
  RockBoulder = 'Rock Boulder',
  RockOutcrop = 'Rock Outcrop',
  Clay = 'Clay',
  Sand = 'Sand',
  Salt = 'Salt',
  OreVein = 'Ore Vein',
  MetalOutcrop = 'Metal Outcrop',
  OceanFishSchool = 'Ocean Fish School',
  LakeFishSchool = 'Lake Fish School',
  ChummedOceanFishSchool = 'Chummed Ocean Fish School',
  Baitfish = 'Baitfish',
  Mollusks = 'Mollusks',
  MonsterDen = 'Monster Den',
  WonderResource = 'Wonder Resource',
  EnergyFont = 'Energy Font',
  Research = 'Research',
  Note = 'Note',
  Bones = 'Bones',
  Door = 'Door',
  Obstacle = 'Obstacle',
  DepletedResource = 'Depleted Resource'
}